requirements.txt
pytorch==1.9.0
matplotlib==3.3.2
numpy==1.18.5
scipy==1.5.4
scikit-learn==0.23.2
mat73==0.46
mne==0.23.4
seaborn==0.11.2
pandas==1.3.4
statannotations==0.4.4
clint
requests